bqp wrapper functions placed into biqbin_bqp.py from biqbin_base.py
